package renamed, simplified wsgi
