+++ powertoy/static/style.css
